Improvements to block batch reader (to allow computing models by block batch)
