+++ src/op_analytics/datapipeline/etl/loadbq/superchain_raw.py
@@ -7,9 +7,6 @@ from op_analytics.coreutils.partitioned.location import DataLocation
 from op_analytics.coreutils.partitioned.paths import get_dt, get_root_path
 from op_analytics.coreutils.partitioned.reader import DataReader
 from op_analytics.datapipeline.chains.goldsky_chains import goldsky_mainnet_chains
-from op_analytics.datapipeline.etl.ingestion.markers import (
-    INGESTION_MARKERS_TABLE,
-)
 from op_analytics.datapipeline.etl.ingestion.reader_bydate import construct_readers_bydate
 
 from .loader import bq_load
@@ -21,13 +18,6 @@ log = structlog.get_logger()
 MARKERS_TABLE = "superchain_raw_bigquery_markers"
 
 BQ_PUBLIC_DATASET = "superchain_raw"
-
-INGESTION_ROOT_PATHS = [
-    "ingestion/blocks_v1",
-    "ingestion/transactions_v1",
-    "ingestion/logs_v1",
-    "ingestion/traces_v1",
-]
 
 
 @bound_contextvars(pipeline_step="load_superchain_raw_to_bq")
@@ -48,8 +38,6 @@ def load_superchain_raw_to_bq(
         chains=chains,
         range_spec=range_spec,
         read_from=DataLocation.GCS,
-        markers_table=INGESTION_MARKERS_TABLE,
-        root_paths=INGESTION_ROOT_PATHS,
     )
 
     if dryrun:
