+++ tests/op_datasets/etl/ingestion/test_reader.py
@@ -7,6 +7,7 @@ from op_analytics.datapipeline.etl.ingestion.reader_bydate import are_inputs_rea
 
 MARKER_PATHS_DATA = [
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-22/000011400000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -16,6 +17,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011400000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -25,6 +27,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011420000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -43,6 +46,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-24/000011440000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -52,6 +56,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-24/000011460000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -61,6 +66,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-24/000011480000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -70,6 +76,7 @@ MARKER_PATHS_DATA = [
         "num_blocks": 20000,
     },
     {
+        "marker_path": "dummy",
         "chain": "fraxtal",
         "data_path": "ingestion/traces_v1/chain=fraxtal/dt=2024-10-24/000011500000.parquet",
         "root_path": "ingestion/traces_v1",
@@ -85,7 +92,7 @@ def test_are_inputs_ready():
     markers_df = pl.DataFrame(MARKER_PATHS_DATA, schema_overrides={"num_blocks": pl.Int32})
     dateval = datetime.date(2024, 10, 23)
 
-    is_ready, actual = are_inputs_ready(
+    input_data = are_inputs_ready(
         markers_df=markers_df,
         dateval=dateval,
         root_paths_to_check={
@@ -93,8 +100,8 @@ def test_are_inputs_ready():
         },
         storage_location=DataLocation.GCS,
     )
-    assert is_ready
-    assert actual == {
+    assert input_data.is_complete
+    assert input_data.data_paths == {
         "ingestion/traces_v1": [
             "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011400000.parquet",
             "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011420000.parquet",
@@ -107,7 +114,7 @@ def test_not_ready_01():
     markers_df = pl.DataFrame(MARKER_PATHS_DATA[:-4], schema_overrides={"num_blocks": pl.Int32})
     dateval = datetime.date(2024, 10, 23)
 
-    is_ready, actual = are_inputs_ready(
+    input_data = are_inputs_ready(
         markers_df=markers_df,
         dateval=dateval,
         root_paths_to_check={
@@ -115,21 +122,15 @@ def test_not_ready_01():
         },
         storage_location=DataLocation.GCS,
     )
-    assert not is_ready
-    assert actual == {
-        "ingestion/traces_v1": [
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011400000.parquet",
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011420000.parquet",
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011440000.parquet",
-        ]
-    }
+    assert not input_data.is_complete
+    assert input_data.data_paths is None
 
 
 def test_not_ready_02():
     markers_df = pl.DataFrame(MARKER_PATHS_DATA, schema_overrides={"num_blocks": pl.Int32})
     dateval = datetime.date(2024, 10, 23)
 
-    is_ready, actual = are_inputs_ready(
+    input_data = are_inputs_ready(
         markers_df=markers_df,
         dateval=dateval,
         root_paths_to_check={
@@ -138,12 +139,5 @@ def test_not_ready_02():
         },
         storage_location=DataLocation.GCS,
     )
-    assert not is_ready
-    assert actual == {
-        "ingestion/logs_v1": [],
-        "ingestion/traces_v1": [
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011400000.parquet",
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011420000.parquet",
-            "gs://oplabs-tools-data-sink/ingestion/traces_v1/chain=fraxtal/dt=2024-10-23/000011440000.parquet",
-        ],
-    }
+    assert not input_data.is_complete
+    assert input_data.data_paths is None
